--- script/dockerversion
@@ -1,12 +1,0 @@
-// AUTOGENERATED FILE; see /go/src/github.com/docker/docker/hack/make/.go-autogen
-package dockerversion
-
-var (
-	GITCOMMIT string = "traefik-import"
-	VERSION   string = "traefik-import"
-	BUILDTIME string = "traefik-import"
-
-	IAMSTATIC string = "traefik-import"
-	INITSHA1  string = "traefik-import"
-	INITPATH  string = "traefik-import"
-)
